backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
chromadb==0.4.18
numpy<2.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
pypdf2==3.0.1
groq>=0.37.0
python-dateutil==2.8.2
apscheduler==3.10.4
python-docx==1.1.0
